requirements.txt
crewai==0.60.0
pymed==0.8.9
pandas==2.1.4
matplotlib==3.8.2
python-dotenv==1.0.0
langchain-ollama==0.1.1
streamlit==1.38.0
crewai-tools==0.8.3
biopython==1.84
ollama==0.3.3
